refactor: clean up link-role-misused rule
