+++ src/rules/link-role-misused/index.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Text } from 'react-native';
-import rule from './link-role-misused';
+import rule from '.';
 import AccessibilityEngine from 'react-native-accessibility-engine';
 
 const run = (component: React.ReactElement<any>) => {
