+++ src/rules/link-role-misused/index.ts
@@ -1,2 +1,23 @@
-import linkRoleMisused from './link-role-misused';
-export default linkRoleMisused;
+import type { Rule } from '../../types';
+import { isText } from '../../helpers';
+
+const rule: Rule = {
+  id: 'accessibility-role-link-misused',
+  matcher: (node) => isText(node.type),
+  assertion: (node) => {
+    const { onPress, accessibilityRole } = node.props;
+    if (!onPress) {
+      return accessibilityRole !== 'link';
+    }
+    return true;
+  },
+  help: {
+    problem:
+      "The 'link' accessibility role should only be used when text is clickable",
+    solution:
+      "Set the 'accessibilityRole' prop to 'text' or add an 'onPress' prop",
+    link: '',
+  },
+};
+
+export default rule;
